Add files via upload
Kaunde Designs is an online clothing store that can be developed using HTML, CSS, and Django Python for the backend. HTML and CSS will be used to create the overall structure, look, and feel of the website. HTML and CSS will allow for the creation of an intuitive and user-friendly interface, allowing customers to quickly find the items they need and make purchases with ease. Additionally, Django Python can be used for the backend to power the website. Django provides a robust and secure framework for creating and managing web applications. It provides the necessary tools to support the development of Kaunde Designs, including authentication and authorization, database access, and other features.